[refactor] cartUtils 디렉토리 변경
 - hooks/utils에서 /utils로
