+++ src/basic/__tests__/basic.test.tsx
@@ -9,7 +9,7 @@ import {
 import { useState } from 'react';
 import { describe, expect, test } from 'vitest';
 import { useCart, useCoupons, useProducts } from '../../refactoring/hooks';
-import * as cartUtils from '../../refactoring/hooks/utils/cartUtils';
+import * as cartUtils from '../../refactoring/utils/cartUtils';
 import { AdminPage } from '../../refactoring/views/AdminPage';
 import { CartPage } from '../../refactoring/views/CartPage';
 import { CartItem, Coupon, Product } from '../../types';
